inference/requirements.txt
PyMuPDF==1.21.1
transformers==4.18.0
textract==1.6.5
python-doctr[torch]==0.6.0
python-magic==0.4.25

torch==1.9.1+cu111 
torchvision==0.10.1+cu111 
torchaudio==0.9.1 
-f https://download.pytorch.org/whl/torch_stable.html

